Structure crate into submodules and add README
* New modules: atomic, iter, ops, util
* Tests for atomic moved into the atomic submodule
* Small fixes for documentation
* README.md
