+++ README.md
@@ -1,0 +1,20 @@
+# hibitset
+
+Provides hierarchical bit sets, which allow very fast iteration on 
+sparse data structures.
+
+## Usage
+
+Just add this to your `Cargo.toml`:
+
+```toml
+[dependencies]
+hibitset = "0.1"
+```
+
+## License
+
+This library is licensed under the Apache License 2.0,
+see [the LICENSE file][li] for more information.
+
+[li]: LICENSE
